Refactor. Fix :bug: calculating widths.
Refactor the getItemPosition method.
Fix `_getOuterWidth` for Firefox because offsetWidth/Height return
integers for Firefox. Now it uses `getBoundingClientRect` to determine
the width.
Up the column threshold from `0.03` to `0.3` because Safari (desktop &
mobile) return integers for the client rect and offsetWidth.
Add a note about the native Android browser with 4.1.0 and 4.1.1 where
it incorrectly thinks it supports unprefixed transitions.